backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
aiomysql>=0.2.0
bcrypt>=4.0.0,<5
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
httpx>=0.27.0
asyncio-mqtt>=0.16.0
# paho-mqtt 2.x 移除 message_retry_set，与 asyncio-mqtt 不兼容，须锁定 1.x
paho-mqtt>=1.6.1,<2
pydantic-settings>=2.2.0
